requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
pyyaml
python-multipart